Add implementation and tests for gax::Status (#4)
* Add implementation and tests for gax::Status

The gax::Status type is intended to be a cornerstone of the client
library. It will be a part of the eventual StatusOr template type,
and is used to determine retryability of failed operations.

The type is default constructable, compares equality, defines
operator<<, and provides convenience query methods for its component attributes.

The underlying StatusCode is intended to closely resemble
grpc::StatusCode, the semantics of which are documented in:
https://grpc.io/grpc/cpp/classgrpc_1_1_status.html

Includes tweaks to RetryPolicy and associated tests to remove
templating and to instead concretely depend on gax::Status.