Add tests and impl for StatusOr<T>
StatusOr is a template type that wraps a value, usually a
protobuf/grpc message, or a Status indicating why an operation failed.

StatusOr is copyable and movable and constructable from Status or T by
const or rvalue reference. It provides overloads for operators *, ->,
and bool() with similar semantics to a smart pointer.
If StatusOr is constructed from a Status or copied/moved from a
StatusOr that was construced from a Status, the underlying T is _not_
constructed or destructed.

The semantics are similar to Rust's Result<T,E>

Constructing a StatusOr from an OK status is not allowed: it results
in a spurious free on value_ when the StatusOr is destructed. Instead
of allowing this behavior, StatusOr(Status) checks and aborts if this
precondition is violated.

Similarly, calling value() if !ok() generates an abort. This is
different from invoking operator*, which results in Undefined Behavior
when !ok().