OperationsStub and Operation<ResultT, MetadataT> (#33)
* Basic tests, intf, and impl for OperationsStub and Operation<ResultT, MetadataT>

OperationsStub is a gax-owned abstract class, similar to the GAPIC
stub, that defines three rpc wrappers: GetOperation, DeleteOperation,
and CancelOperation. GAPIC stubs for clients that define long running
operations and include operations.proto will inherit from
OperationsStub and provide overrides for the three methods.

OperationsClient is a gax-owned concrete class, similar to the GAPIC client, that is responsible for updating, deleting, or canceling an operation.

Operation<ResultT, MetadataT> is a user-visible primitive for
interacting with long-running operations. It is a type-safe wrapper around google::longrunning::Operation.

In the future, an alternative, simpler interface that hides the
details of gax::Operation will be exposed that returns a future object.

A separate review will add a method to GAPIC clients that returns an
OperationsClient sharing the same underlying stub.

A separate review will use protobuf annotations to detect client methods that should return gax::Operation and what the types of its template parameters should be.