First pass tests and impl for PageResult
In order to support multiple use cases and extreme customizations,
not to mention making a more modular, well-designed,
and helpful implementation, paginated method results will be based on three classes
and corresponding iterators.

This is the commit for PageResult, which is a gax-owned type that provides a consistent interface
over individual result pages.