Fix *_gapic_library linking problem (#56)
GAPIC generated client libraries must be statically linked into applications that use them, which is now enforced in the GAPIC bazel rule. `cc_library` generates an empty `.so` file with no symbols or machine code. This could be because the GAPIC bazel rule uses a directory as input instead of a concrete file list.
This could be a general bug in Bazel's native `cc_library` rule implementation. Fixing such a bug is outside the scope of GAPIC, so we rely on a workaround.