Cleanup integration with googleapis repository (relying on latest chagnes in googleapis).
Also removing dependency from api-common-protos to googleapis (since we already depend on it, otherwise we  depend on two copies of same thing).

It looks like api-common-protos is not going anywhere and (there is no current realistic plan to switch to api-common-protos in foreseeable future, and gapic-generator-cpp is the only user of that, while googleapis remains our source of truth). In case if it changes, we can put api-common-protos back easily.