Cleanup integration with googleapis repository (#60)
Also replace dependency on `api-common-protos` with `googleapis` (since we already depend on it, otherwise we depend on two copies of same thing).

It looks like `api-common-protos` is not going anywhere (there is no current realistic plan to switch to `api-common-protos` in foreseeable future, and `gapic-generator-cpp` is the only user of that, while `googleapis` remains our source of truth). In case if it changes, we can put `api-common-protos` back easily.